# pkg/models/user.go
package models

// User model info
// @Description User account information
// @Description with user id and username
type User struct {
	// ID this is userid
	ID            int
	Name          string
	Email         string
	FavoriteColor string
}